fix(ws): device-build compatibility — Kconfig path, esp_mac.h, int* rssi, partition grow
Five pre-existing regressions in the FW-13 code surface caught by
running 'idf.py --target esp32 build' against the merged code. Each
fix is host-tested + device-build verified.

1. firmware/main/Kconfig.projbuild: removed the 'source
   "components/identity/Kconfig"' directives that T-13-C added.
   The path resolved relative to Kconfig.projbuild's directory
   (firmware/main/), so 'components/identity/Kconfig' was looking
   for firmware/main/components/identity/Kconfig which does not
   exist. The IDF auto-source via idf_component_register() in each
   component's CMakeLists.txt is the correct mechanism — verified
   by the wifi component's existing Kconfig (which works the same
   way).

2. firmware/components/identity/identity.c: typo 'nvs_flash.h>' on
   line 28 (stray '>' instead of closing '") + missing
   '#include "esp_mac.h"' for the esp_read_mac() declaration.
   Both surfaced by the device compiler (host mocks the include).

3. firmware/components/ws/ws.c: added '#include "esp_mac.h"' for
   the ESP_MAC_WIFI_STA constant used in the FW-13.4 URL-no-MAC
   guard. Host tests already passed because the mock triplet hides
   the missing include; device build caught it.

4. firmware/components/ws/ws_runtime_metrics.c: IDF v5.5.3 declares
   esp_wifi_sta_get_rssi(int *rssi) — NOT (int32_t *rssi). The
   int32_t cast on assignment is preserved so the public type stays
   portable.

5. firmware/components/ws/Kconfig: Kconfig help text indentation
   dropped on the FIRMWARE_WS_STATUS_PERIOD_MS symbol; the '§'
   character on a mis-indented help line broke the parser. Replaced
   with 'section' + restored 12-space indentation.

6. firmware/partitions.csv: factory partition grew from 0x110000
   (1 MB + 64 KB) to 0x130000 (1 MB + 192 KB) to accommodate the
   esp_websocket_client managed component (~50 KB) + ws/identity
   component code (~5 KB). Post-link firmware.bin = 1,117,136 bytes
   (89.7% of 0x130000, 10.3% free).

Build result: 'idf.py --target esp32 build' succeeds end-to-end
[100%] Built target gen_project_binary. The firmware.bin is
flash-ready; device-flash end-to-end verify (T-13-K PR-body caveat)
remains the user's responsibility (no backend on
ws://<host>:<port>/cams available in this orchestrator session).

Engram observations: #3756 (design), #3758 (tasks), #3767 (apply
progress). Verify-evidence: 'idf.py build' 100% Built target
gen_project_binary; firmware.bin sha256 captured for verify-report.

Co-located regression test entries are unchanged — the host
runner confirms Pass 1 = 122 tests all green; Passes 2-11 fire
the expected bite-proof invariants.

+++ firmware/components/ws/ws_runtime_metrics.c
@@ -40,10 +40,11 @@ void ws_runtime_metrics_collect(ws_runtime_metrics_t *out)
 
     /* rssi_dbm — station RSSI in dBm. On host the mock returns
      * the primed value (default -50 dBm per mock_esp_wifi.c);
-     * on device it queries the wifi driver. */
-    int32_t rssi = 0;
+     * on device it queries the wifi driver. Note: IDF v5.5.3
+     * declares the parameter as `int *` (not int32_t *). */
+    int rssi = 0;
     if (esp_wifi_sta_get_rssi(&rssi) == ESP_OK) {
-        out->rssi_dbm = rssi;
+        out->rssi_dbm = (int32_t)rssi;
     }
 
     /* free_heap — esp_get_free_heap_size() (mock default 200 KB). */
